First attempt at working dynamic outputs
These changes are UNTESTED as I don't have the hardware at this point. Similar approach to the tested output_i2s.cpp / .h, only initialising hardware once and forcing silence out if the object is nominally non-existent (the static parts still exist and service the DMA interrupts).

This will stop crashes, but swapping back and forth between lots of different hardware will probably exhaust the DMA channels eventually. The correct solution would be to create a proper base class for the audio DMA, and make its destructors work correctly. There's a lot of inconsistent but very similar code going on in here which I don't propose to fight with at this point...