First pass at Teensy 3.x compatibility
A number of objects had conditional compilation which was untested when compiling for Teensy 4.x, and didn't work. These now compile OK, though aren't tested at the time of this commit.

Also some fixes related to the OSCAudio library, which needed to be made aware of which objects aren't implemented in Teensy 3.x. in general these consist of ensuring the .h files don't define their guard macros, which in turn leads to the OSCAudio library omitting the objects from its lists. A bit fragile and tricksy, but there you go....