Make AudioOutputSPDIF3 sync to AudioInputSPDIF3 to avoid glitches
AudioOutputSPDIF3 can use the Teensy audio clock at 44.1kHz, OR, if an AudioInputSPDIF3 object (which derives its clock from the incoming audio stream) is instantiated, it can and must be synchronised to that. This approach allows a glitch-free system using SPDIF I/O only.

The design GUI has been modified to allow this combination to pass conflict checks; a better implementation may be required.