Add ability to startFrom with an AudioPreload object
This falls back to a simple file play() if the startFrom parameter is past the end of the pre-load buffer.
Also added some timing check capability so we can see how long it takes from play() to first update, and when the file buffering completes. As expected latency is <= 2.9ms with the pre-load, can be more playing direct from the file; though as it's only taking 0.66-1.08ms to get the file buffered, the difference is minor with the test ode as-is.