Add instrumentation to playback and record classes
Also a couple of instrumented versions of the demo files. An Instrument class simply takes a new value and stores the last, highest and lowest values, counts the number of updates, and provides a way to read and reset these. For the buffered playback and record, we instrument the filesystem read or write time (in microseconds), and the buffer level at the time of read / write (in bytes). We can thus get an idea of the required buffer size and how fast the filesystem is.