Preliminary attempt to make playback work safely even if initiated from ISR
If any play(file) method, or adjustHeaderInfo(), is called from interrupt, e.g. an IntervalTimer callback, the request is deferred and an event queued, so it should start on the next yield(). This will result in extra latency, depending on how often loop() / yield() / delay() get called by the application. However, as before, a play(preload) will start on the next audio update, and is the preferred choice for low-latency applications.

AudioRecordWAVbuffered has *not* been updated in this commit, we'll see if any gremlins are found and when AudioPlayWAVbuffered seems OK I'll do recording.