Implement drive; start implementing internal filtering
External feedback doesn't work, because of the processing delays - we have to take it inside the object, for this example by deriving from the biquad filter (may not be the best choice...). AudioFilterBiquad modifications are a beginning attempt to filter part of a block at a time, because for higher frequencies (> 344.5Hz) we have fewer than 128 samples per cycle.

TBD: unroll the processPart() loop a bit more, so we can cope with an odd sample count. What happens if we're accessing 32-bit entities on a non-multiple-of-4 boundary?