Change exported AudioConnection initalizers to brace form
This allows creation of a class using the exported
code pretty much* as-is, which the old syntax did not.
The class can then be instantiated multiple times
without writing the same topology out many times over,
which is a maintenance nightmare.

See the PlaySynthMusic-class example, and compare
with the PlaySynthMusic one.

* you will probably need to remove any input or
 output nodes, which the Design Tool insists on,
 but should only be instantiated once.