Store designs in separate locations
If you have multiple copies of the Design Tool, they
all shared the same local storage for your design, which
could result in broken topology if e.g. some objects
weren't implemented in one of the branches.

This change uses the URL of the tool instance as
part of the storage key, so every copy should get
a separate record in local storage. Note that this
will not work for web-based tools which don't have
this change imported.