First go at weighted filtering
As discussed at https://forum.pjrc.com/index.php?threads/on-adding-an-internal-synthesizer-engine-to-the-the-mosaichord-midi-controller-using-the-teensy-audio-library.77464/post-363478

The code appears to work:
* use the optional `weight`` parameter to `setFeedbackLevel()``
* much of the class is temporarily public for debug purposes
* feedback level can be >1.0 for now (it's not a good idea, though)

Some pitch inaccuracy when weight is not 0.5, I think because the
effective delay is N+weight samples, and the existing algorithm
therefore has a -0.5 built into it. This could probably be fixed but...

...I'm not finding any significant difference in decay time resulting
from changes to weight, so not sure there's merit in further effort.